refactor: 검색 결과 페이지 추상화
